requirements.txt
#
# Dependencies for mysql-fabric.
#
--allow-external mysql-connector-python
mysql-connector-python
docker-py
python-novaclient
